+++ src/clj_kondo/impl/analyzer.clj
@@ -687,18 +687,18 @@
                 (when (identical? :quote t)
                   (findings/reg-finding!
                    ctx
-                   (node->line (:filename ctx) dupe :quoted-case-test-constant
+                   (node->line (:filename ctx) dupe :case-quoted-test
                                "Case test is compile time constant and should not be quoted.")))
                 (when (symbol? (:value constant))
                   (findings/reg-finding!
                    ctx
-                   (node->line (:filename ctx) dupe :symbol-case-test-constant
-                               "Case test symbol is never evaluated. Mark case with #_{:clj-kondo/ignore [:symbol-case-test-constant]} to get rid of warning."))))
+                   (node->line (:filename ctx) dupe :case-symbol-test
+                               "Case test symbol is compile time constant and is never evaluated."))))
               (let [s-dupe (str dupe)]
                 (when (seen-local s-dupe)
                   (findings/reg-finding!
                    ctx
-                   (node->line (:filename ctx) dupe :duplicate-case-test-constant
+                   (node->line (:filename ctx) dupe :case-duplicate-test
                                (format "Duplicate case test constant: %s" s-dupe))))
                 (when (seq more)
                   (recur more (conj seen-local s-dupe)))))